backend/app/core/config.py
from pydantic_settings import BaseSettings
from pathlib import Path
from dotenv import load_dotenv
import os

# Load from backend/.env
env_path = Path(__file__).resolve().parent.parent.parent / ".env"
load_dotenv(dotenv_path=env_path)

# Also load from project root .env (for MLflow credentials if they are there)
root_env_path = Path(__file__).resolve().parent.parent.parent.parent / ".env"
if root_env_path.exists():
    load_dotenv(dotenv_path=root_env_path)

class Settings(BaseSettings):
    PROJECT_NAME: str = "YouTube Comment Analysis API"
    VERSION: str = "1.0.0"
    API_V1_STR: str = "/api/v1"
    
    # YouTube Data API
    YOUTUBE_API_KEY: str = os.getenv("YOUTUBE_API_KEY", "")
    
    # Model Paths
    BASE_DIR: Path = Path(__file__).resolve().parent.parent.parent.parent
    MODELS_DIR: Path = BASE_DIR / "models"
    
    class Config:
        case_sensitive = True

settings = Settings()
