requirements.txt
loguru
mkdocs
pip
python-dotenv
ruff
tqdm
typer
textacy
pandas
numpy
scikit-learn
seaborn
nltk
gensim
kaggle
emoji
ipykernel
transformers==4.39.3
sentencepiece
torch
langdetect
tqdm
mlflow
pyarrow
emoji
kagglehub
wordcloud
imblearn
xgboost
catboost
optuna
lightgbm
dagshub
pyyaml
dvc
dagshub
fastapi
pytest
uvicorn
google-api-python-client
pydantic-settings
-e .
